Refactor to less fragile canary approach
... and remove some dead frontend code
... and cleanup English in new UI
